// adapters/mtproto/src/mtproto/gen/api.d.ts
export namespace mt {
  export type ResPQ<
    K extends keyof _ResPQ = keyof _ResPQ
  > = ToUnderscore<_ResPQ, K>;
  export type _ResPQ = {
    "mt.resPQ": {
      nonce: Uint8Array;                      // int128
      server_nonce: Uint8Array;               // int128
      pq: Uint8Array;                         // bytes
      server_public_key_fingerprints: bigint[]; // Vector<long>
    },
  };
// ...
  export type Server_DH_Params<
    K extends keyof _Server_DH_Params = keyof _Server_DH_Params
  > = ToUnderscore<_Server_DH_Params, K>;
  export type _Server_DH_Params = {
    "mt.server_DH_params_fail": {
      nonce: Uint8Array;                      // int128
      server_nonce: Uint8Array;               // int128
      new_nonce_hash: Uint8Array;             // int128
    },
    "mt.server_DH_params_ok": {
      nonce: Uint8Array;                      // int128
      server_nonce: Uint8Array;               // int128
      encrypted_answer: Uint8Array;           // bytes
    },
  };
// ...
  export type Set_client_DH_params_answer<
    K extends keyof _Set_client_DH_params_answer = keyof _Set_client_DH_params_answer
  > = ToUnderscore<_Set_client_DH_params_answer, K>;
  export type _Set_client_DH_params_answer = {
    "mt.dh_gen_ok": {
      nonce: Uint8Array;                      // int128
      server_nonce: Uint8Array;               // int128
      new_nonce_hash1: Uint8Array;            // int128
    },
    "mt.dh_gen_retry": {
      nonce: Uint8Array;                      // int128
      server_nonce: Uint8Array;               // int128
      new_nonce_hash2: Uint8Array;            // int128
    },
    "mt.dh_gen_fail": {
      nonce: Uint8Array;                      // int128
      server_nonce: Uint8Array;               // int128
      new_nonce_hash3: Uint8Array;            // int128
    },
  };
// ...
  export type RpcDropAnswer<
    K extends keyof _RpcDropAnswer = keyof _RpcDropAnswer
  > = ToUnderscore<_RpcDropAnswer, K>;
  export type _RpcDropAnswer = {
    "mt.rpc_answer_unknown": {}
    "mt.rpc_answer_dropped_running": {}
    "mt.rpc_answer_dropped": {
      msg_id: bigint;                         // long
      seq_no: number;                         // int
      bytes: number;                          // int
    },
  };
// ...
  export type FutureSalts<
    K extends keyof _FutureSalts = keyof _FutureSalts
  > = ToUnderscore<_FutureSalts, K>;
  export type _FutureSalts = {
    "mt.future_salts": {
      req_msg_id: bigint;                     // long
      now: number;                            // int
      salts: mt.FutureSalt[];                 // vector<mt.FutureSalt>
    },
  };

  export const future_salts: TLConstructor<_FutureSalts, "mt.future_salts">;
  export type Pong<
    K extends keyof _Pong = keyof _Pong
  > = ToUnderscore<_Pong, K>;
  export type _Pong = {
    "mt.pong": {
      msg_id: bigint;                         // long
      ping_id: bigint;                        // long
    },
  };
// ...
  export type DestroyAuthKeyRes<
    K extends keyof _DestroyAuthKeyRes = keyof _DestroyAuthKeyRes
  > = ToUnderscore<_DestroyAuthKeyRes, K>;
  export type _DestroyAuthKeyRes = {
    "mt.destroy_auth_key_ok": {}
    "mt.destroy_auth_key_none": {}
    "mt.destroy_auth_key_fail": {}
  };

  export const destroy_auth_key_ok: TLConstructorEmpty<"mt.destroy_auth_key_ok">;
  export const destroy_auth_key_none: TLConstructorEmpty<"mt.destroy_auth_key_none">;
  export const destroy_auth_key_fail: TLConstructorEmpty<"mt.destroy_auth_key_fail">;
  export type DestroySessionRes<
    K extends keyof _DestroySessionRes = keyof _DestroySessionRes
  > = ToUnderscore<_DestroySessionRes, K>;
  export type _DestroySessionRes = {
    "mt.destroy_session_ok": {
      session_id: bigint;                     // long
    },
    "mt.destroy_session_none": {
      session_id: bigint;                     // long
    },
  };

  export const destroy_session_ok: TLConstructor<_DestroySessionRes, "mt.destroy_session_ok">;
  export const destroy_session_none: TLConstructor<_DestroySessionRes, "mt.destroy_session_none">;
}

export namespace api {
// ...
  export type _Updates = {
    "updatesTooLong": {}
    "updateShortMessage": {
      out?: true;                             // flags.1?true
      mentioned?: true;                       // flags.4?true
      media_unread?: true;                    // flags.5?true
      silent?: true;                          // flags.13?true
      id: number;                             // int
      user_id: bigint;                        // long
      message: string;                        // string
      pts: number;                            // int
      pts_count: number;                      // int
      date: number;                           // int
      fwd_from?: api.MessageFwdHeader;        // flags.2?MessageFwdHeader
      via_bot_id?: bigint;                    // flags.11?long
      reply_to?: api.MessageReplyHeader;      // flags.3?MessageReplyHeader
      entities?: api.MessageEntity[];         // flags.7?Vector<MessageEntity>
      ttl_period?: number;                    // flags.25?int
    },
    "updateShortChatMessage": {
      out?: true;                             // flags.1?true
      mentioned?: true;                       // flags.4?true
      media_unread?: true;                    // flags.5?true
      silent?: true;                          // flags.13?true
      id: number;                             // int
      from_id: bigint;                        // long
      chat_id: bigint;                        // long
      message: string;                        // string
      pts: number;                            // int
      pts_count: number;                      // int
      date: number;                           // int
      fwd_from?: api.MessageFwdHeader;        // flags.2?MessageFwdHeader
      via_bot_id?: bigint;                    // flags.11?long
      reply_to?: api.MessageReplyHeader;      // flags.3?MessageReplyHeader
      entities?: api.MessageEntity[];         // flags.7?Vector<MessageEntity>
      ttl_period?: number;                    // flags.25?int
    },
    "updateShort": {
      update: api.Update;                     // Update
      date: number;                           // int
    },
    "updatesCombined": {
      updates: api.Update[];                  // Vector<Update>
      users: api.User[];                      // Vector<User>
      chats: api.Chat[];                      // Vector<Chat>
      date: number;                           // int
      seq_start: number;                      // int
      seq: number;                            // int
    },
    "updates": {
      updates: api.Update[];                  // Vector<Update>
      users: api.User[];                      // Vector<User>
      chats: api.Chat[];                      // Vector<Chat>
      date: number;                           // int
      seq: number;                            // int
    },
    "updateShortSentMessage": {
      out?: true;                             // flags.1?true
      id: number;                             // int
      pts: number;                            // int
      pts_count: number;                      // int
      date: number;                           // int
      media?: api.MessageMedia;               // flags.9?MessageMedia
      entities?: api.MessageEntity[];         // flags.7?Vector<MessageEntity>
      ttl_period?: number;                    // flags.25?int
    },
  };
// ...
  export const updateShortSentMessage: TLConstructor<_Updates, "updateShortSentMessage">;
// ...
  export type InputClientProxy<
    K extends keyof _InputClientProxy = keyof _InputClientProxy
  > = ToUnderscore<_InputClientProxy, K>;
  export type _InputClientProxy = {
    "inputClientProxy": {
      address: string;                        // string
      port: number;                           // int
    },
  };
// ...
  export type JSONValue<
    K extends keyof _JSONValue = keyof _JSONValue
  > = ToUnderscore<_JSONValue, K>;
  export type _JSONValue = {
    "jsonNull": {}
    "jsonBool": {
      value: boolean;                         // Bool
    },
    "jsonNumber": {
      value: number;                          // double
    },
    "jsonString": {
      value: string;                          // string
    },
    "jsonArray": {
      value: api.JSONValue[];                 // Vector<JSONValue>
    },
    "jsonObject": {
      value: api.JSONObjectValue[];           // Vector<JSONObjectValue>
    },
  };
// ...
}
// ...
export namespace mt {
  export const req_pq: TLMethod<{
    nonce: Uint8Array                       // int128
  }, ResPQ>
  export const req_pq_multi: TLMethod<{
    nonce: Uint8Array                       // int128
  }, ResPQ>
  export const req_DH_params: TLMethod<{
    nonce: Uint8Array                       // int128
    server_nonce: Uint8Array                // int128
    p: Uint8Array                           // bytes
    q: Uint8Array                           // bytes
    public_key_fingerprint: bigint          // long
    encrypted_data: Uint8Array              // bytes
  }, Server_DH_Params>
  export const set_client_DH_params: TLMethod<{
    nonce: Uint8Array                       // int128
    server_nonce: Uint8Array                // int128
    encrypted_data: Uint8Array              // bytes
  }, Set_client_DH_params_answer>
  export const rpc_drop_answer: TLMethod<{
    req_msg_id: bigint                      // long
  }, RpcDropAnswer>
  export const get_future_salts: TLMethod<{
    num: number                             // int
  }, FutureSalts>
  export const ping: TLMethod<{
    ping_id: bigint                         // long
  }, Pong>
  export const ping_delay_disconnect: TLMethod<{
    ping_id: bigint                         // long
    disconnect_delay: number                // int
  }, Pong>
  export const http_wait: TLMethod<{
    max_delay: number                       // int
    wait_after: number                      // int
    max_wait: number                        // int
  }, unknown>
  export const destroy_auth_key: TLMethod<void, DestroyAuthKeyRes>
  export const destroy_session: TLMethod<{
    session_id: bigint                      // long
  }, DestroySessionRes>
}

export const invokeAfterMsg: TLApiMethod<"invokeAfterMsg", {
  msg_id: bigint                          // long
  query: any                              // !X
}, any>
export const invokeAfterMsgs: TLApiMethod<"invokeAfterMsgs", {
  msg_ids: bigint[]                       // Vector<long>
  query: any                              // !X
}, any>
export const initConnection: TLApiMethod<"initConnection", {
  api_id: number                          // int
  device_model: string                    // string
  system_version: string                  // string
  app_version: string                     // string
  system_lang_code: string                // string
  lang_pack: string                       // string
  lang_code: string                       // string
  proxy?: api.InputClientProxy            // flags.0?InputClientProxy
  params?: api.JSONValue                  // flags.1?JSONValue
  query: any                              // !X
}, any>
export const invokeWithLayer: TLApiMethod<"invokeWithLayer", {
  layer: number                           // int
  query: any                              // !X
}, any>
export const invokeWithoutUpdates: TLApiMethod<"invokeWithoutUpdates", {
  query: any                              // !X
}, any>